+++ src/ecp.c
@@ -603,7 +603,7 @@ EcpSetCommand(int ac, char *av[], void *arg)
 
   if (ac == 0)
     return(-1);
-  switch ((int) arg)
+  switch ((intptr_t)arg)
   {
     case SET_KEY:
       if (ac != 1)
